+++ server/test/integration/BraveAlerterConfiguratorTest/getAlertSessionBySessionIdAndAlertApiKeyTest.js
@@ -15,7 +15,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSessionBySessio
     this.expectedChatbotState = CHATBOT_STATE.WAITING_FOR_CATEGORY
     this.expectedIncidentCategory = 'No One Inside'
     this.expectedLocationDisplayName = 'TEST LOCATION'
-    this.expectedLocationPhoneNumber = '+17772225555'
+    this.expectedRespondedByPhoneNumber = '+17772225555'
     this.expectedTwilioPhoneNumber = '+3336661234'
     this.expectedIncidentCategoryKeys = ['1', '2', '3']
     this.expectedIncidentCategories = ['No One Inside', 'Person responded', 'None of the above']
@@ -23,7 +23,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSessionBySessio
 
     // Insert a location in the DB
     const client = await factories.clientDBFactory(db, {
-      responderPhoneNumber: this.expectedLocationPhoneNumber,
+      responderPhoneNumbers: [this.expectedRespondedByPhoneNumber],
       alertApiKey: this.alertApiKey,
       incidentCategories: this.expectedIncidentCategories,
     })
@@ -39,7 +39,7 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSessionBySessio
       alertType: ALERT_TYPE.SENSOR_DURATION,
       chatbotState: this.expectedChatbotState,
       incidentCategory: this.expectedIncidentCategory,
-      respondedByPhoneNumber: this.expectedLocationPhoneNumber,
+      respondedByPhoneNumber: this.expectedRespondedByPhoneNumber,
     })
   })
 
@@ -54,9 +54,9 @@ describe('BraveAlerterConfigurator.js integration tests: getAlertSessionBySessio
     const expectedAlertSession = new AlertSession(
       this.session.id,
       this.expectedChatbotState,
-      this.expectedLocationPhoneNumber,
+      this.expectedRespondedByPhoneNumber,
       this.expectedIncidentCategory,
-      [this.expectedLocationPhoneNumber],
+      [this.expectedRespondedByPhoneNumber],
       this.expectedIncidentCategoryKeys,
       this.expectedIncidentCategories,
     )